+++ package.json
@@ -11,8 +11,10 @@
   "dependencies": {
     "expo": "~54.0.2",
     "expo-status-bar": "~3.0.8",
+    "lottie-react-native": "^7.3.4",
     "react": "19.1.0",
-    "react-native": "0.81.4"
+    "react-native": "0.81.4",
+    "react-native-safe-area-context": "^5.6.1"
   },
   "devDependencies": {
     "@types/react": "~19.1.0",
