+++ src/assets/colors/index.ts
@@ -1,0 +1,11 @@
+const COLORS = {
+  bg: "#DDF3EA", // minty background
+  card: "#E9F7EF", // light green card
+  text: "#20323A", // dark slate
+  sub: "#3E5963",
+  button: "#F35C4A", // coral/red
+  buttonText: "#FFFFFF",
+  border: "#B7E1CF",
+};
+
+export default COLORS;
